Version 2020.04.30 Changes
- Added clipboard support to the Linux version
- Increase max local notifications to 100
- Fixed CopyNetworkMessage failing to copy messages that were not fully read
- Copied network messages can now be read and added to
- Added AddNetworkMessageBye and GetNetworkMessageByte
- Added an error if the Mod command is used with a divisor of 0 to avoid a crash
- Added SetSoundDeviceMode for iOS to choose between exclusive audio allowing audio from background apps to continue
- iOS now defaults to allowing audio from background apps to continue
- iOS Tier 2 projects have been updated to use launch storyboards instead of images
- iOS export has been updated to use a launch logo instead of a splash image, the logo will be centered on screen in any orientation
- Fixed SetShadowCascades not applying the given values
- Updated Android to target API 29 as required by Google
- Added FixCameraToObject command
- Added GetCameraWorldX/Y/Z commands to get the world coordinates of cameras attached to objects
- Fixed SetObjectReceiveShadow being set to 0 still overwriting object texture stages with shadow maps
- Fixed a crash when running a Spriter animation that had a different number of keyframes per varying value
- Spriter image pivots can now be animated
- Spriter animations now support negative scales
- Spriter files no longer need to have the same bone structure across all their animations
- Numerous other small fixes to improve Spriter support
- Removed Facebook SDK in favour of the ShareText/ShareImage commands
- Updated AdMob SDK to version 7.58.0 on iOS
- Updated Chartboost SDK to version 8.1.0 on iOS
- Updated AmazonAds SDK to version 3.0.0 on iOS
- Updated Firebase SDK to version 6.22.0 on iOS
- Android apps resuming from background will now resume faster

+++ zlib/unzip.c
@@ -188,8 +188,15 @@ typedef struct
     int isZip64;
 
 #    ifndef NOUNCRYPT
-    unsigned long keys[3];     /* keys defining the pseudo-random sequence */
-    const unsigned long* pcrc_32_tab;
+
+	//PE: Support arm 64 bit. should with with all platforms, 21-02-2020: need testing.
+	unsigned int keys[3];     /* keys defining the pseudo-random sequence */
+	const unsigned int* pcrc_32_tab;
+
+	//PE: Old code.
+	//unsigned long keys[3];     /* keys defining the pseudo-random sequence */
+	//const unsigned long* pcrc_32_tab;
+
 #    endif
 } unz64_s;
